backend/requirements.txt
fastapi===0.115.12
ultralytics==8.3.78
opencv-python==4.11.0.86
moviepy===1.0.3
uvicorn===0.22.0
python-multipart===0.0.20
websockets==12.0
pillow==10.3.0
transformers==4.38.0
sentencepiece===0.2.0
onnxruntime===1.22.1